Refactor BasePage: move special locatorrs and methods to special pages; Add ResetPasswordPage
